usefull
git-svn-id: svn://svn.dd-wrt.com/DD-WRT@38992 52c4871e-980c-0410-b1e0-e73912ce01f8

+++ src/router/rules/e2fsprogs.mk
@@ -13,7 +13,7 @@ e2fsprogs-clean:
 e2fsprogs-install:
 	-make -C e2fsprogs install DESTDIR=$(INSTALLDIR)/e2fsprogs
 
-#ifneq ($(CONFIG_E2FSPROGS_ADV),y)
+ifneq ($(CONFIG_E2FSPROGS_ADV),y)
 	-rm -f $(INSTALLDIR)/e2fsprogs/sbin/*fsck*
 	rm -rf $(INSTALLDIR)/e2fsprogs/usr/bin
 	rm -rf $(INSTALLDIR)/e2fsprogs/usr/sbin
@@ -32,7 +32,7 @@ e2fsprogs-install:
 	rm -f $(INSTALLDIR)/e2fsprogs/sbin/fsck*
 	rm -f $(INSTALLDIR)/e2fsprogs/lib/libss*
 	rm -f $(INSTALLDIR)/e2fsprogs/lib/e2initrd_helper	
-#endif
+endif
 	rm -rf $(INSTALLDIR)/e2fsprogs/usr/share
 	rm -rf $(INSTALLDIR)/e2fsprogs/usr/man
 	rm -rf $(INSTALLDIR)/e2fsprogs/usr
